requirements.txt
mysql-connector-python 
deep-translator